+++ changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.16.0
+
+### Fixed
+
+- Run embedding models on CPU to work around memory copy issue
+
 ## 0.15.0 - 2024-01-04
 
 ### Changed